Fix deleting site with multilang page and content
1. Clean logic for the case of a call of deleteContents from site deletion.
We only need to do the check for unarchived content in other langs to prevent the
deletion  in cases different than site deletion.

2. Remove call to indexAPI.removeContentFromIndex in the old confusing logic for the
case of site deletion since it is called for every content later in the method

3. Remove deletion of ContentVersionInfo inside the loop that deals with multitrees.
This is because if a parent page and child content were in the list being iterated
and the parent page was first in that list, it would remove its version info, thus
when getting content child's multitree, it would be impossible to find the parent by
identifier since its version info was removed in the previous iteration. The deletion
all ContentVersionInfo ocurrs later in the method.

Commit for #7561